# Conflicts:
#   RaceOverlay/MainWindow.xaml.cs

+++ RaceOverlay/Data/Mapper.cs
@@ -128,7 +128,7 @@ public class Mapper
         data.LocalCarTelemetry.BrakeBias = irsdkSharper.Data.GetFloat("dcBrakeBias");
         try
         {
-            data.LocalCarTelemetry.Tc1 = irsdkSharper.Data.GetInt("dcTractionControl");
+            data.LocalCarTelemetry.Tc1 = irsdkSharper.Data.GetFloat("dcTractionControl");
         }
         catch (Exception e)
         {
@@ -137,7 +137,7 @@ public class Mapper
 
         try
         {
-            data.LocalCarTelemetry.Tc2 = irsdkSharper.Data.GetInt("dcTractionControl2");
+            data.LocalCarTelemetry.Tc2 = irsdkSharper.Data.GetFloat("dcTractionControl2");
         }
         catch (Exception e)
         {
@@ -146,7 +146,7 @@ public class Mapper
 
         try
         {
-            data.LocalCarTelemetry.Abs = irsdkSharper.Data.GetInt("dcABS");
+            data.LocalCarTelemetry.Abs = irsdkSharper.Data.GetFloat("dcABS");
         }
         catch (Exception e)
         {
@@ -176,4 +176,5 @@ public class Mapper
         // Return Dataset
         return data;
     }
+    
 }
